requirements.txt
python-dotenv==1.0.0
selenium==4.14.0
undetected_chromedriver==3.5.3
